+++ cache-dashboard/src/main/java/com/saucesubfresh/cache/admin/entity/OpenCacheMetricsDO.java
@@ -5,6 +5,7 @@ import com.baomidou.mybatisplus.annotation.TableName;
 import lombok.Data;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 /**
@@ -41,27 +42,27 @@ public class OpenCacheMetricsDO implements Serializable {
 	/**
 	 * 总请求总数
 	 */
-	private long requestCount;
+	private Long requestCount;
 
 	/**
 	 * 总命中总数
 	 */
-	private long hitCount;
+	private Long hitCount;
 
 	/**
 	 * 总未命中总数
 	 */
-	private long missCount;
+	private Long missCount;
 
 	/**
 	 * 命中率
 	 */
-	private double hitRate;
+	private BigDecimal hitRate;
 
 	/**
 	 * 未命中率
 	 */
-	private double missRate;
+	private BigDecimal missRate;
 
 	/**
 	 * 任务日志创建时间
